Remove DocumentsTab and MongoDocumentsTab. Fix build issues.
